[Store] Split SSD unit tests
