+++ mooncake-store/tests/ha/snapshot/master_service/ssd/lifecycle_test.cpp
@@ -1,0 +1,287 @@
+#include "fixture.h"
+
+namespace mooncake::test {
+TEST_F(MasterServiceSSDSnapshotTest, PutEndBothReplica) {
+    CreateMasterServiceWithSSDFeat("/mnt/ssd");
+
+    constexpr size_t buffer = 0x300000000;
+    constexpr size_t size = 1024 * 1024 * 64;
+    std::string segment_name = "test_segment";
+    Segment segment;
+    segment.id = generate_uuid();
+    segment.name = segment_name;
+    segment.base = buffer;
+    segment.size = size;
+    segment.te_endpoint = segment.name;
+    UUID client_id = generate_uuid();
+
+    auto mount_result = service_->MountSegment(segment, client_id);
+    ASSERT_TRUE(mount_result.has_value());
+
+    std::string key = "disk_key";
+    uint64_t slice_length = 1024;
+    ReplicateConfig config;
+    config.replica_num = 1;
+
+    auto put_start_result =
+        service_->PutStart(client_id, key, TenantId::Default(), slice_length, config);
+    ASSERT_TRUE(put_start_result.has_value());
+    auto replicas = put_start_result.value();
+    ASSERT_EQ(2, replicas.size());
+
+    bool has_mem = false, has_disk = false;
+    for (const auto& r : replicas) {
+        if (r.is_memory_replica()) has_mem = true;
+        if (r.is_disk_replica()) has_disk = true;
+    }
+    EXPECT_TRUE(has_mem);
+    EXPECT_TRUE(has_disk);
+
+    auto get_result = service_->GetReplicaList(key, TenantId::Default());
+    ASSERT_FALSE(get_result.has_value());
+    EXPECT_EQ(ErrorCode::REPLICA_IS_NOT_READY, get_result.error());
+
+    // PutEnd for both memory and disk
+    EXPECT_TRUE(service_->PutEnd(client_id, key, TenantId::Default(), ReplicaType::MEMORY)
+                    .has_value());
+    EXPECT_TRUE(service_->PutEnd(client_id, key, TenantId::Default(), ReplicaType::DISK)
+                    .has_value());
+
+    get_result = service_->GetReplicaList(key, TenantId::Default());
+    ASSERT_TRUE(get_result.has_value());
+    EXPECT_EQ(2, get_result.value().replicas.size());
+
+    for (const auto& r : get_result.value().replicas) {
+        EXPECT_EQ(ReplicaStatus::COMPLETE, r.status);
+    }
+}
+
+TEST_F(MasterServiceSSDSnapshotTest, PutRevokeDiskReplica) {
+    CreateMasterServiceWithSSDFeat("/mnt/ssd");
+
+    constexpr size_t buffer = 0x300000000;
+    constexpr size_t size = 1024 * 1024 * 64;
+    std::string segment_name = "test_segment";
+    Segment segment;
+    segment.id = generate_uuid();
+    segment.name = segment_name;
+    segment.base = buffer;
+    segment.size = size;
+    segment.te_endpoint = segment.name;
+    UUID client_id = generate_uuid();
+
+    auto mount_result = service_->MountSegment(segment, client_id);
+    ASSERT_TRUE(mount_result.has_value());
+
+    std::string key = "revoke_key";
+    uint64_t slice_length = 1024;
+    ReplicateConfig config;
+    config.replica_num = 1;
+
+    ASSERT_TRUE(
+        service_->PutStart(client_id, key, TenantId::Default(), slice_length, config)
+            .has_value());
+    EXPECT_TRUE(service_->PutEnd(client_id, key, TenantId::Default(), ReplicaType::MEMORY)
+                    .has_value());
+
+    auto get_result = service_->GetReplicaList(key, TenantId::Default());
+    ASSERT_TRUE(get_result.has_value());
+    EXPECT_EQ(1, get_result.value().replicas.size());
+    ASSERT_TRUE(get_result.value().replicas[0].is_memory_replica());
+
+    EXPECT_TRUE(
+        service_->PutRevoke(client_id, key, TenantId::Default(), ReplicaType::DISK)
+            .has_value());
+
+    get_result = service_->GetReplicaList(key, TenantId::Default());
+    ASSERT_TRUE(get_result.has_value());
+    EXPECT_EQ(1, get_result.value().replicas.size());
+    ASSERT_TRUE(get_result.value().replicas[0].is_memory_replica());
+}
+
+TEST_F(MasterServiceSSDSnapshotTest, PutRevokeMemoryReplica) {
+    CreateMasterServiceWithSSDFeat("/mnt/ssd");
+
+    constexpr size_t buffer = 0x300000000;
+    constexpr size_t size = 1024 * 1024 * 64;
+    std::string segment_name = "test_segment";
+    Segment segment;
+    segment.id = generate_uuid();
+    segment.name = segment_name;
+    segment.base = buffer;
+    segment.size = size;
+    segment.te_endpoint = segment.name;
+    UUID client_id = generate_uuid();
+
+    auto mount_result = service_->MountSegment(segment, client_id);
+    ASSERT_TRUE(mount_result.has_value());
+
+    std::string key = "revoke_key";
+    uint64_t slice_length = 1024;
+    ReplicateConfig config;
+    config.replica_num = 1;
+
+    ASSERT_TRUE(
+        service_->PutStart(client_id, key, TenantId::Default(), slice_length, config)
+            .has_value());
+    EXPECT_TRUE(
+        service_->PutRevoke(client_id, key, TenantId::Default(), ReplicaType::MEMORY)
+            .has_value());
+
+    auto get_result = service_->GetReplicaList(key, TenantId::Default());
+    ASSERT_FALSE(get_result.has_value());
+    EXPECT_EQ(ErrorCode::REPLICA_IS_NOT_READY, get_result.error());
+
+    EXPECT_TRUE(service_->PutEnd(client_id, key, TenantId::Default(), ReplicaType::DISK)
+                    .has_value());
+    get_result = service_->GetReplicaList(key, TenantId::Default());
+    ASSERT_TRUE(get_result.has_value());
+    EXPECT_EQ(1, get_result.value().replicas.size());
+    ASSERT_TRUE(get_result.value().replicas[0].is_disk_replica());
+}
+
+TEST_F(MasterServiceSSDSnapshotTest, PutRevokeBothReplica) {
+    CreateMasterServiceWithSSDFeat("/mnt/ssd");
+
+    constexpr size_t buffer = 0x300000000;
+    constexpr size_t size = 1024 * 1024 * 64;
+    std::string segment_name = "test_segment";
+    Segment segment;
+    segment.id = generate_uuid();
+    segment.name = segment_name;
+    segment.base = buffer;
+    segment.size = size;
+    segment.te_endpoint = segment.name;
+    UUID client_id = generate_uuid();
+
+    auto mount_result = service_->MountSegment(segment, client_id);
+    ASSERT_TRUE(mount_result.has_value());
+
+    std::string key = "revoke_key";
+    uint64_t slice_length = 1024;
+    ReplicateConfig config;
+    config.replica_num = 1;
+
+    ASSERT_TRUE(
+        service_->PutStart(client_id, key, TenantId::Default(), slice_length, config)
+            .has_value());
+    EXPECT_TRUE(
+        service_->PutRevoke(client_id, key, TenantId::Default(), ReplicaType::DISK)
+            .has_value());
+
+    auto get_result = service_->GetReplicaList(key, TenantId::Default());
+    ASSERT_FALSE(get_result.has_value());
+    EXPECT_EQ(ErrorCode::REPLICA_IS_NOT_READY, get_result.error());
+
+    EXPECT_TRUE(
+        service_->PutRevoke(client_id, key, TenantId::Default(), ReplicaType::MEMORY)
+            .has_value());
+    get_result = service_->GetReplicaList(key, TenantId::Default());
+    ASSERT_FALSE(get_result.has_value());
+    EXPECT_EQ(ErrorCode::OBJECT_NOT_FOUND, get_result.error());
+}
+
+TEST_F(MasterServiceSSDSnapshotTest, RemoveKey) {
+    CreateMasterServiceWithSSDFeat("/mnt/ssd");
+
+    constexpr size_t buffer = 0x300000000;
+    constexpr size_t size = 1024 * 1024 * 64;
+    std::string segment_name = "test_segment";
+    Segment segment;
+    segment.id = generate_uuid();
+    segment.name = segment_name;
+    segment.base = buffer;
+    segment.size = size;
+    segment.te_endpoint = segment.name;
+    UUID client_id = generate_uuid();
+
+    auto mount_result = service_->MountSegment(segment, client_id);
+    ASSERT_TRUE(mount_result.has_value());
+
+    std::string key = "remove_key";
+    uint64_t slice_length = 1024;
+    ReplicateConfig config;
+    config.replica_num = 1;
+
+    ASSERT_TRUE(
+        service_->PutStart(client_id, key, TenantId::Default(), slice_length, config)
+            .has_value());
+    EXPECT_TRUE(service_->PutEnd(client_id, key, TenantId::Default(), ReplicaType::MEMORY)
+                    .has_value());
+    EXPECT_TRUE(service_->PutEnd(client_id, key, TenantId::Default(), ReplicaType::DISK)
+                    .has_value());
+
+    EXPECT_TRUE(service_->Remove(key, TenantId::Default()).has_value());
+
+    auto get_result = service_->GetReplicaList(key, TenantId::Default());
+    EXPECT_FALSE(get_result.has_value());
+    EXPECT_EQ(ErrorCode::OBJECT_NOT_FOUND, get_result.error());
+}
+
+TEST_F(MasterServiceSSDSnapshotTest, EvictObject) {
+    // Keep the lease expiry wait below the client live TTL. This test verifies
+    // eviction and snapshot restore, not the process-wide default lease TTL.
+    constexpr uint64_t kv_lease_ttl = 2000;
+    auto service_config = MasterServiceConfig::builder()
+                              .set_root_fs_dir("/mnt/ssd")
+                              .set_default_kv_lease_ttl(kv_lease_ttl)
+                              .build();
+    CreateMasterServiceWithSSDFeatAndConfig(service_config);
+    // Mount a segment that can hold about 1024 * 16 objects.
+    // As the eviction is processed separately for each shard,
+    // we need to fill each shard with enough objects to thoroughly
+    // test the eviction process.
+    constexpr size_t buffer = 0x300000000;
+    constexpr size_t size = 1024 * 1024 * 16 * 15;
+    constexpr size_t object_size = 1024 * 15;
+    std::string segment_name = "test_segment";
+    Segment segment;
+    segment.id = generate_uuid();
+    segment.name = segment_name;
+    segment.base = buffer;
+    segment.size = size;
+    segment.te_endpoint = segment.name;
+    UUID client_id = generate_uuid();
+    auto mount_result = service_->MountSegment(segment, client_id);
+    ASSERT_TRUE(mount_result.has_value());
+
+    // Verify if we can put objects more than the segment can hold
+    int success_puts = 0;
+    for (int i = 0; i < 1024 * 16 + 50; ++i) {
+        std::string key = "test_key" + std::to_string(i);
+        uint64_t slice_length = object_size;
+        ReplicateConfig config;
+        config.replica_num = 1;
+        auto put_start_result =
+            service_->PutStart(client_id, key, TenantId::Default(), slice_length, config);
+        if (put_start_result.has_value()) {
+            auto put_end_mem_result = service_->PutEnd(
+                client_id, key, TenantId::Default(), ReplicaType::MEMORY);
+            auto put_end_disk_result =
+                service_->PutEnd(client_id, key, TenantId::Default(), ReplicaType::DISK);
+            ASSERT_TRUE(put_end_mem_result.has_value());
+            ASSERT_TRUE(put_end_disk_result.has_value());
+            success_puts++;
+        } else {
+            // wait for eviction to work
+            std::this_thread::sleep_for(std::chrono::milliseconds(50));
+        }
+    }
+    ASSERT_GT(success_puts, 1024 * 16);
+
+    // Verify if we can get objects more than the segment can hold
+    int success_gets = 0;
+    for (int i = 0; i < 1024 * 16 + 50; ++i) {
+        std::string key = "test_key" + std::to_string(i);
+        auto get_result = service_->GetReplicaList(key, TenantId::Default());
+        if (get_result.has_value()) {
+            success_gets++;
+        }
+    }
+    ASSERT_GT(success_gets, 1024 * 16);
+
+    std::this_thread::sleep_for(std::chrono::milliseconds(kv_lease_ttl));
+    // service_->RemoveAll();
+}
+
+}  // namespace mooncake::test
