+++ mooncake-store/tests/master_service/ssd/expiration_test.cpp
@@ -1,0 +1,115 @@
+#include "fixture.h"
+
+namespace mooncake::test {
+TEST_F(MasterServiceSSDTest, PutStartExpires) {
+    // Reset storage space metrics.
+    MasterMetricManager::instance().reset_allocated_mem_size();
+    MasterMetricManager::instance().reset_total_mem_capacity();
+
+    MasterServiceConfig master_config;
+    master_config.root_fs_dir = "/mnt/ssd";
+    master_config.put_start_discard_timeout_sec = 3;
+    master_config.put_start_release_timeout_sec = 5;
+    std::unique_ptr<MasterService> service_(new MasterService(master_config));
+
+    constexpr size_t kReplicaCnt = 2;  // 1 memory replica + 1 disk replica
+    constexpr size_t kBaseAddr = 0x300000000;
+    constexpr size_t kSegmentSize = 1024 * 1024 * 16;  // 16MB
+
+    // Mount a segment.
+    std::string segment_name = "test_segment";
+    Segment segment;
+    segment.id = generate_uuid();
+    segment.name = segment_name;
+    segment.base = kBaseAddr;
+    segment.size = kSegmentSize;
+    segment.te_endpoint = segment.name;
+    auto client_id = generate_uuid();
+    auto mount_result = service_->MountSegment(segment, client_id);
+    ASSERT_TRUE(mount_result.has_value());
+
+    std::string key = "test_key";
+    uint64_t value_length = 16 * 1024 * 1024;  // 16MB
+    uint64_t slice_length = value_length;
+    ReplicateConfig config;
+
+    auto test_discard_replica = [&](ReplicaType discard_type) {
+        const auto reserve_type = discard_type == ReplicaType::MEMORY
+                                      ? ReplicaType::DISK
+                                      : ReplicaType::MEMORY;
+
+        // Put key, should success.
+        auto put_start_result =
+            service_->PutStart(client_id, key, TenantId::Default(), slice_length, config);
+        EXPECT_TRUE(put_start_result.has_value());
+        auto replica_list = put_start_result.value();
+        EXPECT_EQ(replica_list.size(), kReplicaCnt);
+        for (size_t i = 0; i < kReplicaCnt; i++) {
+            EXPECT_EQ(ReplicaStatus::PROCESSING, replica_list[i].status);
+        }
+
+        // Complete the reserved replica.
+        auto put_end_result =
+            service_->PutEnd(client_id, key, TenantId::Default(), reserve_type);
+        EXPECT_TRUE(put_end_result.has_value());
+
+        // Wait for a while until the put-start expired.
+        for (size_t i = 0; i <= master_config.put_start_discard_timeout_sec;
+             i++) {
+            // Keep mounted segments alive.
+            auto result = service_->Ping(client_id);
+            EXPECT_TRUE(result.has_value());
+            // Protect the key from eviction.
+            auto get_result = service_->GetReplicaList(key, TenantId::Default());
+            EXPECT_TRUE(get_result.has_value());
+            std::this_thread::sleep_for(std::chrono::seconds(1));
+        }
+
+        // Put key again, should fail because the object has had an completed
+        // replica.
+        put_start_result =
+            service_->PutStart(client_id, key, TenantId::Default(), slice_length, config);
+        EXPECT_FALSE(put_start_result.has_value());
+        EXPECT_EQ(put_start_result.error(), ErrorCode::OBJECT_ALREADY_EXISTS);
+
+        // Wait for a while until the discarded replicas are released.
+        for (size_t i = 0; i <= master_config.put_start_release_timeout_sec;
+             i++) {
+            // Keep mounted segments alive.
+            auto result = service_->Ping(client_id);
+            EXPECT_TRUE(result.has_value());
+            // Protect the key from eviction.
+            auto get_result = service_->GetReplicaList(key, TenantId::Default());
+            EXPECT_TRUE(get_result.has_value());
+            std::this_thread::sleep_for(std::chrono::seconds(1));
+        }
+
+        // Try PutEnd the discarded replica.
+        put_end_result =
+            service_->PutEnd(client_id, key, TenantId::Default(), discard_type);
+        EXPECT_TRUE(put_end_result.has_value());
+
+        // Check that the key has only one replica.
+        auto get_result = service_->GetReplicaList(key, TenantId::Default());
+        EXPECT_TRUE(get_result.has_value());
+        EXPECT_EQ(get_result.value().replicas.size(), 1);
+        if (reserve_type == ReplicaType::MEMORY) {
+            EXPECT_TRUE(get_result.value().replicas[0].is_memory_replica());
+        } else {
+            EXPECT_TRUE(get_result.value().replicas[0].is_disk_replica());
+        }
+
+        // Wait for the key to expire.
+        for (size_t i = 0; i <= DEFAULT_DEFAULT_KV_LEASE_TTL / 1000; i++) {
+            auto result = service_->Ping(client_id);
+            EXPECT_TRUE(result.has_value());
+            std::this_thread::sleep_for(std::chrono::seconds(1));
+        }
+        service_->RemoveAll();
+    };
+
+    test_discard_replica(ReplicaType::DISK);
+    test_discard_replica(ReplicaType::MEMORY);
+}
+
+}  // namespace mooncake::test
