+++ mooncake-store/tests/master_service/ssd/fixture.h
@@ -1,0 +1,95 @@
+#pragma once
+
+// Shared fixture and helpers for MasterService SSD tests.
+
+#include "master_service.h"
+
+#include <glog/logging.h>
+#include <gtest/gtest.h>
+
+#include <memory>
+#include <thread>
+#include <vector>
+
+#include "master_metric_manager.h"
+#include "types.h"
+
+namespace mooncake::test {
+
+std::unique_ptr<MasterService> CreateMasterServiceWithSSDFeat(
+    const std::string& root_fs_dir) {
+    return std::make_unique<MasterService>(
+        MasterServiceConfig::builder().set_root_fs_dir(root_fs_dir).build());
+}
+
+class MasterServiceSSDTest : public ::testing::Test {
+   protected:
+    void SetUp() override {
+        google::InitGoogleLogging("MasterServiceTest");
+        FLAGS_logtostderr = true;
+    }
+
+    void TearDown() override { google::ShutdownGoogleLogging(); }
+};
+
+std::unique_ptr<MasterService> CreateSsdAwareOffloadService() {
+    MasterServiceConfig config;
+    config.enable_offload = true;
+    config.default_kv_lease_ttl = 0;
+    config.allocation_strategy_type =
+        AllocationStrategyType::SSD_FREE_RATIO_FIRST;
+    return std::make_unique<MasterService>(config);
+}
+
+void MountMemoryAndLocalDisk(MasterService& service, const UUID& client_id,
+                             const std::string& segment_name,
+                             size_t base_addr) {
+    Segment segment;
+    segment.id = generate_uuid();
+    segment.name = segment_name;
+    segment.base = base_addr;
+    segment.size = 64 * 1024 * 1024;
+    segment.te_endpoint = segment.name;
+
+    ASSERT_TRUE(service.MountSegment(segment, client_id).has_value());
+    ASSERT_TRUE(service.MountLocalDiskSegment(client_id, true).has_value());
+    ASSERT_TRUE(service.ReportSsdCapacity(client_id, 1000).has_value());
+}
+
+void PutAndOffload(MasterService& service, const UUID& client_id,
+                   const std::string& key, int64_t object_size,
+                   const std::string& local_disk_endpoint) {
+    ReplicateConfig config;
+    config.replica_num = 1;
+
+    ASSERT_TRUE(service.PutStart(client_id, key, TenantId::Default(), object_size, config)
+                    .has_value());
+    ASSERT_TRUE(service.PutEnd(client_id, key, TenantId::Default(), ReplicaType::MEMORY)
+                    .has_value());
+
+    StorageObjectMetadata metadata;
+    metadata.data_size = object_size;
+    metadata.transport_endpoint = local_disk_endpoint;
+    OffloadTaskItem task{
+        .tenant_id = TenantId::Default().value(), .key = key, .size = object_size};
+    ASSERT_TRUE(service.NotifyOffloadSuccess(client_id, {task}, {metadata})
+                    .has_value());
+}
+
+void ExpectNextAllocationOnSegment(MasterService& service,
+                                   const UUID& client_id,
+                                   const std::string& key,
+                                   const std::string& expected_segment) {
+    ReplicateConfig config;
+    config.replica_num = 1;
+    auto result = service.PutStart(client_id, key, TenantId::Default(), 64, config);
+    ASSERT_TRUE(result.has_value());
+    ASSERT_EQ(result->size(), 1u);
+    ASSERT_TRUE((*result)[0].is_memory_replica());
+    EXPECT_EQ((*result)[0]
+                  .get_memory_descriptor()
+                  .buffer_descriptor.transport_endpoint_,
+              expected_segment);
+}
+
+}  // namespace mooncake::test
